Added repo console test and split up the console tests into 3 files
